Folded BioXASShuttersGroup into BioXASShutters. Modified BioXASSideBeamline and BioXASMainBeamline to use the new utilities class in BioXASBeamline for the shutters. #1967
