+++ bioxasCommon.pri
@@ -126,7 +126,6 @@ HEADERS += \
     source/ui/BioXAS/BioXASMirrorButton.h \
     $$PWD/source/ui/BioXAS/BioXASButtonEditorBar.h \
 	$$PWD/source/ui/BioXAS/BioXASBeamStatusBar.h \
-	source/beamline/BioXAS/BioXASShuttersGroup.h \
 	source/beamline/BioXAS/BioXASZebraTimeSeconds.h \
 	source/beamline/BioXAS/BioXASZebraSoftInputControl.h \
     source/beamline/BioXAS/BioXASZebraLogicBlock.h \
@@ -257,7 +256,6 @@ SOURCES += \
     source/ui/BioXAS/BioXASMirrorButton.cpp \
     $$PWD/source/ui/BioXAS/BioXASButtonEditorBar.cpp \
 	$$PWD/source/ui/BioXAS/BioXASBeamStatusBar.cpp \
-	source/beamline/BioXAS/BioXASShuttersGroup.cpp \
 	source/beamline/BioXAS/BioXASZebraTimeSeconds.cpp \
 	source/beamline/BioXAS/BioXASZebraSoftInputControl.cpp \
     source/beamline/BioXAS/BioXASZebraLogicBlock.cpp \
