+++ source/beamline/BioXAS/BioXASMainBeamline.h
@@ -24,8 +24,6 @@ along with Acquaman.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "beamline/BioXAS/BioXASBeamline.h"
 
-#include "beamline/BioXAS/BioXASShutters.h"
-#include "beamline/BioXAS/BioXASBeamStatus.h"
 #include "beamline/BioXAS/BioXASMainM1Mirror.h"
 #include "beamline/BioXAS/BioXASMainMonochromator.h"
 #include "beamline/BioXAS/BioXASMainM2Mirror.h"
@@ -68,11 +66,6 @@ public:
 	virtual BioXASMainMonochromator *mono() const { return mono_; }
 	/// Returns the beamline M2 mirror.
 	virtual BioXASM2Mirror *m2Mirror() const { return m2Mirror_; }
-
-	/// Returns the endstation safety shutter.
-	virtual CLSExclusiveStatesControl* endstationShutter() const { return endstationShutter_; }
-	/// Returns the shutters.
-	virtual BioXASShutters* shutters() const { return shutters_; }
 
 	/// Returns the beam status.
 	virtual BioXASBeamStatus* beamStatus() const { return beamStatus_; }
@@ -146,11 +139,6 @@ protected:
 	/// The M2 mirror.
 	BioXASMainM2Mirror *m2Mirror_;
 
-	/// The endstation shutter.
-	CLSExclusiveStatesControl *endstationShutter_;
-	/// The shutters.
-	BioXASShutters *shutters_;
-
 	/// The beam status.
 	BioXASBeamStatus *beamStatus_;
 
@@ -166,9 +154,6 @@ protected:
 	BioXASMainCryostatStage *cryostatStage_;
 	/// Endstation table
 	BioXASEndstationTable *endstationTable_;
-
-	/// Utilities
-	BioXASBeamlineUtilities *utilities_;
 
 	// Scaler controls
 	/// Scaler
