+++ source/beamline/BioXAS/BioXASShutters.cpp
@@ -7,10 +7,10 @@
 BioXASShutters::BioXASShutters(const QString &name, QObject *parent) :
 	BioXASShuttersGroup(name, parent)
 {
-	// Initialize class variables.
+	// Setup basic value options.
 
-	frontEndShutters_ = 0;
-	endstationShutter_ = 0;
+	addOption(Open, "Open");
+	addOption(Closed, "Closed", true);
 }
 
 BioXASShutters::~BioXASShutters()
@@ -18,57 +18,119 @@ BioXASShutters::~BioXASShutters()
 
 }
 
+bool BioXASShutters::isOpen() const
+{
+	bool result = false;
+
+	if (isConnected() && areAllChildrenState1())
+		result = true;
+
+	return result;
+}
+
+bool BioXASShutters::isClosed() const
+{
+	bool result = false;
+
+	if (isConnected() && areAnyChildrenState2())
+		result = true;
+
+	return result;
+}
+
 bool BioXASShutters::hasShutter(AMControl *control) const
 {
 	return hasChildControl(control);
 }
 
-void BioXASShutters::setFrontEndShutters(BioXASFrontEndShutters *newShutters)
+void BioXASShutters::addShutter(AMControl *newShutter, double openValue, double closedValue)
 {
-	if (frontEndShutters_ != newShutters) {
-
-		if (frontEndShutters_)
-			removeShutter(frontEndShutters_);
-
-		frontEndShutters_ = newShutters;
-
-		if (frontEndShutters_)
-			addShutter(frontEndShutters_, BioXASFrontEndShutters::Open, BioXASFrontEndShutters::Closed);
-
-		emit frontEndShuttersChanged(frontEndShutters_);
-	}
+	if (addBiStateControl(newShutter, openValue, closedValue))
+		emit shuttersChanged();
 }
 
-void BioXASShutters::setEndstationShutter(CLSExclusiveStatesControl *newShutter)
+void BioXASShutters::removeShutter(AMControl *shutter)
 {
-	if (endstationShutter_ != newShutter) {
+	if (removeBiStateControl(shutter))
+		emit shuttersChanged();
+}
 
-		if (endstationShutter_)
-			removeShutter(endstationShutter_);
+void BioXASShutters::clearShutters()
+{
+	if (clearBiStateControls())
+		emit shuttersChanged();
+}
 
-		endstationShutter_ = newShutter;
+AMAction3* BioXASShutters::createMoveAction(double setpoint)
+{
+	AMAction3 *result = 0;
 
-		if (endstationShutter_)
-			addShutter(endstationShutter_, CLSExclusiveStatesControl::Open, CLSExclusiveStatesControl::Closed);
-
-		emit endstationShutterChanged(endstationShutter_);
+	switch (int(setpoint)) {
+	case Open:
+		result = createMoveToOpenAction();
+		break;
+	case Closed:
+		result = createMoveToClosedAction();
+		break;
+	default:
+		break;
 	}
+
+	return result;
 }
 
 AMAction3* BioXASShutters::createMoveToOpenAction()
 {
-	AMListAction3 *action = new AMListAction3(new AMListActionInfo3("Opening shutters", "Opening shutters"), AMListAction3::Sequential);
-	action->addSubAction(createMoveChildToOpen(endstationShutter_));
-	action->addSubAction(createMoveChildToOpen(frontEndShutters_));
+	AMAction3 *action = createMoveChildrenToState1Action();
+
+	if (action) {
+		action->info()->setShortDescription("Opening shutters.");
+		action->info()->setLongDescription("Opening shutters.");
+	}
 
 	return action;
 }
 
 AMAction3* BioXASShutters::createMoveToClosedAction()
 {
-	AMListAction3 *action = new AMListAction3(new AMListActionInfo3("Closing shutters", "Closing shutters"), AMListAction3::Sequential);
-	action->addSubAction(createMoveChildToClosed(frontEndShutters_));
-	action->addSubAction(createMoveChildToClosed(endstationShutter_));
+	AMAction3 *action = createMoveChildrenToState2Action();
+
+	if (action) {
+		action->info()->setShortDescription("Closing shutters.");
+		action->info()->setLongDescription("Closing shutters.");
+	}
 
 	return action;
+}
+
+AMAction3* BioXASShutters::createMoveChildToOpen(AMControl *child)
+{
+	return createMoveChildToState1Action(child);
+}
+
+AMAction3* BioXASShutters::createMoveChildToClosed(AMControl *child)
+{
+	return createMoveChildToState2Action(child);
+}
+
+AMAction3* BioXASShutters::createCheckChildIsOpen(AMControl *child, double timeoutSec)
+{
+	return createCheckChildAtState1Action(child, timeoutSec);
+}
+
+AMAction3* BioXASShutters::createCheckChildIsClosed(AMControl *child, double timeoutSec)
+{
+	return createCheckChildAtState2Action(child, timeoutSec);
+}
+
+int BioXASShutters::currentIndex() const
+{
+	int result = enumNames().indexOf("Unknown");
+
+	if (isOpen())
+		result = Open;
+	else if (isClosed())
+		result = Closed;
+
+	return result;
 }
